Minor Fixes

+++ Challenges Java/src/chapter8/challenge3/Main.java
@@ -7,14 +7,14 @@ public class Main {
 	public static void main(String[] args) {
 		try {
 			String witdthS =  JOptionPane.showInputDialog("Insert the Witdth: ");
-			Double width = Double.valueOf(witdthS);
+			float width = Float.valueOf(witdthS);
 			
-			Double heigth = Double.valueOf(JOptionPane.showInputDialog("Insert Heigth"));
-			Double cost = Double.valueOf(JOptionPane.showInputDialog("Insert Cost of Carpet"));
+			float heigth = Float.valueOf(JOptionPane.showInputDialog("Insert Heigth"));
+			float cost = Float.valueOf(JOptionPane.showInputDialog("Insert Cost of Carpet"));
 			
 			RoomDimension roomDim = new RoomDimension(width, heigth);
 			RoomCarpet carpet = new RoomCarpet(roomDim, cost);
-			Double price = carpet.calculatePrice();
+			float price = carpet.calculatePrice();
 			System.out.println("The price is : "+price);
 			JOptionPane.showMessageDialog(null, "The price is : "+ price);
 		}
